@@ -41,14 +41,15 @@
     "uglify"
   ],
   "engines": {
-    "node": ">=12"
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
   },
-  "main": "./src/htmlminifier.js",
-  "module": "./dist/htmlminifier.mjs",
+  "type": "module",
+  "main": "./dist/htmlminifier.cjs",
+  "module": "./src/htmlminifier.js",
   "exports": {
     ".": {
-      "require": "./src/htmlminifier.js",
-      "import": "./dist/htmlminifier.mjs"
+      "require": "./dist/htmlminifier.cjs",
+      "import": "./src/htmlminifier.js"
     },
     "./dist/*": "./dist/*.js",
     "./package.json": "./package.json"
@@ -63,9 +64,9 @@
   ],
   "scripts": {
     "build": "rollup -c",
-    "test:node": "jest --verbose",
-    "test:web": "jest --verbose --environment=jsdom",
-    "test:watch": "jest verbose --watch",
+    "test:node": "NODE_OPTIONS=--experimental-vm-modules jest --verbose",
+    "test:web": "NODE_OPTIONS=--experimental-vm-modules jest --verbose --environment=jsdom",
+    "test:watch": "NODE_OPTIONS=--experimental-vm-modules jest verbose --watch",
     "test": "npm run test:node",
     "serve": "vite",
     "build:docs": "vite build --base /html-minifier-terser/",
@@ -83,12 +84,12 @@
   },
   "devDependencies": {
     "@commitlint/cli": "^15.0.0",
-    "@jest/globals": "^27.4.2",
+    "@jest/globals": "^27.4.4",
     "@rollup/plugin-commonjs": "^21.0.1",
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^13.0.6",
     "alpinejs": "^3.7.0",
-    "commitlint-config-non-conventional": "^1.0.0",
+    "commitlint-config-non-conventional": "^1.0.1",
     "eslint": "^7.32.0",
     "eslint-config-standard": "^16.0.3",
     "eslint-plugin-import": "^2.25.3",
@@ -96,12 +97,12 @@
     "eslint-plugin-promise": "^5.2.0",
     "husky": "^7.0.4",
     "is-ci": "^3.0.1",
-    "jest": "^27.4.3",
+    "jest": "^27.4.4",
     "lint-staged": "^12.1.2",
-    "rollup": "^2.60.2",
+    "rollup": "^2.61.0",
     "rollup-plugin-polyfill-node": "^0.8.0",
     "rollup-plugin-terser": "^7.0.2",
-    "vite": "^2.6.14"
+    "vite": "^2.7.1"
   },
   "benchmarkDependencies": {
     "chalk": "^4.1.2",
